norm lasts details
