+++ parser/dyntree.h
@@ -6,10 +6,9 @@
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/13 18:11:07 by lray              #+#    #+#             */
-/*   Updated: 2023/10/30 13:45:56 by lray             ###   ########.fr       */
+/*   Updated: 2023/10/31 21:02:39 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #ifndef DYNTREE_H
 # define DYNTREE_H
