+++ vars/grpvar_utils.c
@@ -1,21 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   handle_sigint_heredoc.c                            :+:      :+:    :+:   */
+/*   grpvar_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 17:48:33 by lray              #+#    #+#             */
-/*   Updated: 2023/10/31 21:06:12 by lray             ###   ########.fr       */
+/*   Created: 2023/10/31 20:52:24 by lray              #+#    #+#             */
+/*   Updated: 2023/10/31 20:52:46 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-void	handle_sigint_heredoc(int sig)
+void	grpvar_free(t_grpvar *grpvar)
 {
-	(void) sig;
-	g_in_heredoc = 0;
-	ioctl(STDIN_FILENO, TIOCSTI, "\n");
-	rl_on_new_line();
+	if (grpvar)
+	{
+		if (grpvar->global)
+		{
+			lstvar_free(grpvar->global);
+			grpvar->global = NULL;
+		}
+		if (grpvar->local)
+		{
+			lstvar_free(grpvar->local);
+			grpvar->local = NULL;
+		}
+		free(grpvar);
+		grpvar = NULL;
+	}
 }
