+++ vars/lstvar.h
@@ -6,7 +6,7 @@
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 16:10:55 by lray              #+#    #+#             */
-/*   Updated: 2023/10/07 14:52:30 by lray             ###   ########.fr       */
+/*   Updated: 2023/10/31 21:07:54 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,6 +31,5 @@ size_t		lstvar_has(t_lstvar *lstvar, char *name);
 int			lstvar_remove(t_lstvar *lstvar, size_t index);
 void		lstvar_free(t_lstvar *lstvar);
 t_dynarrstr	*lstvar_to_array(t_lstvar *lstvar);
-
 
 #endif
